+++ graphdatascience/graph/graph_entity_ops_runner.py
@@ -87,14 +87,18 @@ class GraphNodePropertiesRunner(GraphEntityOpsBaseRunner):
 
         # new format was requested, but the query was run via Cypher
         if separate_property_columns and "propertyValue" in result.keys():
-            result = result.pivot(index="nodeId", columns="nodeProperty", values="propertyValue")
-            result = result.reset_index()
+            wide_result = result.pivot(index=["nodeId"], columns=["nodeProperty"], values="propertyValue")
+            if "listNodeLabels" in config.keys():
+                # nodeLabels cannot be an index column of the pivot as its not hashable
+                # so we need to manually join it back in
+                labels_df = result[["nodeId", "nodeLabels"]].set_index("nodeId")
+                wide_result = wide_result.join(labels_df, on="nodeId")
+            result = wide_result.reset_index()
             result.columns.name = None
         # old format was requested but the query was run via Arrow
         elif not separate_property_columns and "propertyValue" not in result.keys():
-            result = result.melt(id_vars=["nodeId"]).rename(
-                columns={"variable": "nodeProperty", "value": "propertyValue"}
-            )
+            id_vars = ["nodeId", "nodeLabels"] if config.get("listNodeLabels", False) else ["nodeId"]
+            result = result.melt(id_vars=id_vars).rename(columns={"variable": "nodeProperty", "value": "propertyValue"})
 
         if db_node_properties:
             duplicate_properties = set(db_node_properties).intersection(set(node_properties))