+++ graphdatascience/query_runner/arrow_query_runner.py
@@ -6,7 +6,9 @@ from typing import Any, Dict, List, Optional, Tuple
 
 import pyarrow.flight as flight
 from pandas import DataFrame, Series
+from pyarrow import ChunkedArray, Table, chunked_array
 from pyarrow.flight import ClientMiddleware, ClientMiddlewareFactory
+from pyarrow.types import is_dictionary  # type: ignore
 
 from ..call_parameters import CallParameters
 from ..server_version.server_version import ServerVersion
@@ -117,6 +119,11 @@ class ArrowQueryRunner(QueryRunner):
             property_name = params["properties"]
             node_labels = params["entities"]
 
+            config = {"node_property": property_name, "node_labels": node_labels}
+
+            if "listNodeLabels" in params["config"]:
+                config["list_node_labels"] = params["config"]["listNodeLabels"]
+
             if self._server_version < new_endpoint_server_version:
                 endpoint = "gds.graph.streamNodeProperty"
             else:
@@ -126,15 +133,16 @@ class ArrowQueryRunner(QueryRunner):
                         old_endpoint="gds.graph.streamNodeProperty", new_endpoint="gds.graph.nodeProperty.stream"
                     )
 
-            return self._run_arrow_property_get(
-                graph_name, endpoint, {"node_property": property_name, "node_labels": node_labels}
-            )
+            return self._run_arrow_property_get(graph_name, endpoint, config)
         elif (
             old_endpoint := ("gds.graph.streamNodeProperties" == endpoint)
         ) or "gds.graph.nodeProperties.stream" == endpoint:
             graph_name = params["graph_name"]
-            property_names = params["properties"]
-            node_labels = params["entities"]
+
+            config = {"node_properties": params["properties"], "node_labels": params["entities"]}
+
+            if "listNodeLabels" in params["config"]:
+                config["list_node_labels"] = params["config"]["listNodeLabels"]
 
             if self._server_version < new_endpoint_server_version:
                 endpoint = "gds.graph.streamNodeProperties"
@@ -147,7 +155,7 @@ class ArrowQueryRunner(QueryRunner):
             return self._run_arrow_property_get(
                 graph_name,
                 endpoint,
-                {"node_properties": property_names, "node_labels": node_labels},
+                config,
             )
         elif (
             old_endpoint := ("gds.graph.streamRelationshipProperty" == endpoint)
@@ -267,9 +275,14 @@ class ArrowQueryRunner(QueryRunner):
         ticket = flight.Ticket(json.dumps(payload).encode("utf-8"))
 
         get = self._flight_client.do_get(ticket)
-        result: DataFrame = get.read_pandas()
+        arrow_table = get.read_all()
 
-        return result
+        if configuration.get("list_node_labels", False):
+            # GDS 2.5 had an inconsistent naming of the node labels column
+            new_colum_names = ["nodeLabels" if i == "labels" else i for i in arrow_table.column_names]
+            arrow_table = arrow_table.rename_columns(new_colum_names)
+
+        return self._sanitize_arrow_table(arrow_table).to_pandas()  # type: ignore
 
     def create_graph_constructor(
         self, graph_name: str, concurrency: int, undirected_relationship_types: Optional[List[str]]
@@ -284,6 +297,22 @@ class ArrowQueryRunner(QueryRunner):
         return ArrowGraphConstructor(
             database, graph_name, self._flight_client, concurrency, undirected_relationship_types
         )
+
+    def _sanitize_arrow_table(self, arrow_table: Table) -> Table:
+        dict_encoded_fields = [
+            (idx, field) for idx, field in enumerate(arrow_table.schema) if is_dictionary(field.type)
+        ]
+        for idx, field in dict_encoded_fields:
+            try:
+                field.type.to_pandas_dtype()
+            except NotImplementedError:
+                # we need to decode the dictionary column before transforming to pandas
+                if isinstance(arrow_table[field.name], ChunkedArray):
+                    decoded_col = chunked_array([chunk.dictionary_decode() for chunk in arrow_table[field.name].chunks])
+                else:
+                    decoded_col = arrow_table[field.name].dictionary_decode()
+                arrow_table = arrow_table.set_column(idx, field.name, decoded_col)
+        return arrow_table
 
 
 class AuthFactory(ClientMiddlewareFactory):  # type: ignore