+++ graphdatascience/tests/integration/test_graph_ops.py
@@ -407,6 +407,47 @@ def test_graph_nodeProperties_stream_with_arrow(gds: GraphDataScience) -> None:
     assert {e for e in name_values["propertyValue"]} == {"nodeA", "nodeB", "nodeC"}
 
 
+@pytest.mark.compatible_with(min_inclusive=ServerVersion(2, 5, 0))
+def test_graph_nodeProperties_stream_listNodeLabels_with_arrow(gds: GraphDataScience) -> None:
+    G, _ = gds.graph.project(GRAPH_NAME, {"Node": {"properties": ["x"]}}, "*")
+
+    result = gds.graph.nodeProperties.stream(G, ["x"], concurrency=2, listNodeLabels=True)
+
+    assert list(result.keys()) == ["nodeId", "nodeLabels", "nodeProperty", "propertyValue"]
+
+    x_values = result[result.nodeProperty == "x"]
+    assert {e for e in x_values["propertyValue"]} == {1, 2, 3}
+
+    assert [e for e in result["nodeLabels"]] == [["Node"], ["Node"], ["Node"]]
+
+
+@pytest.mark.compatible_with(min_inclusive=ServerVersion(2, 5, 0))
+def test_graph_nodeProperties_stream_listNodeLabels(gds_without_arrow: GraphDataScience) -> None:
+    G, _ = gds_without_arrow.graph.project(GRAPH_NAME, {"Node": {"properties": ["x"]}}, "*")
+
+    result = gds_without_arrow.graph.nodeProperties.stream(G, ["x"], concurrency=2, listNodeLabels=True)
+
+    assert list(result.keys()) == ["nodeId", "nodeProperty", "propertyValue", "nodeLabels"]
+
+    x_values = result[result.nodeProperty == "x"]
+    assert {e for e in x_values["propertyValue"]} == {1, 2, 3}
+
+    assert [e for e in result["nodeLabels"]] == [["Node"], ["Node"], ["Node"]]
+
+
+@pytest.mark.compatible_with(min_inclusive=ServerVersion(2, 5, 0))
+def test_graph_nodeProperties_stream_listNodeLabels_with_seperate_cols(gds_without_arrow: GraphDataScience) -> None:
+    G, _ = gds_without_arrow.graph.project(GRAPH_NAME, {"Node": {"properties": ["x", "y"]}}, "*")
+
+    result = gds_without_arrow.graph.nodeProperties.stream(
+        G, ["x", "y"], concurrency=2, listNodeLabels=True, separate_property_columns=True
+    )
+
+    assert list(result.keys()) == ["nodeId", "x", "y", "nodeLabels"]
+
+    assert [e for e in result["nodeLabels"]] == [["Node"] for _ in range(6)]
+
+
 def test_graph_streamNodeProperties_with_arrow_separate_property_columns(gds: GraphDataScience) -> None:
     G, _ = gds.graph.project(GRAPH_NAME, {"Node": {"properties": ["x", "y"]}}, "*")